上位机下载配置正常，重启之后读取数据没有数据

+++ RTT_P2P/Project/DistributionComponents/station.h
@@ -24,7 +24,7 @@ extern ErrorCode PacketDecodeStationMessage(AreaID* area, uint8_t* data, uint16_
 extern ErrorCode TransmitEncodeStationMessage_All(StationPoint* point, uint16_t destAddress);
 extern ErrorCode ParseNanopb(StationPoint*, uint16_t sourceAddress, uint8_t* pdata, uint8_t len);
 extern ErrorCode PacketEncodeStationMessage_All(const StationPoint* const point, PointUint8* packet, uint16_t addLen, uint16_t offset);
-extern void  ManagerAddStationByStationMessage(uint8_t data[], uint8_t len, StationManger* manger);
+extern void  ManagerAddStationByStationMessage(uint8_t data[], uint16_t len, StationManger* manger);
 extern ErrorCode PacketDecodeStationMessage_ALL(StationMessage* pMessage, uint8_t* data, uint16_t len);
 #ifdef	__cplusplus
 }
